files converter/requirements.txt
flask
werkzeug
pdf2image
Pillow
python-docx
pytesseract
fpdf
